+++ CHANGELOG.md
@@ -8,3 +8,4 @@ All notable changes CombineExt will be documented in this file.
 * Add `Zip` for multiple Publishers (#6)
 * Add `Nwise` for consecutive elements (#27)
 * Add `Merge` extension on `Collection`
+* Add `ScanInto` and `ReduceInto` (#13)